modified:   functions/addpreference.func.php 	modified:   pages/clientmain.php 	modified:   pages/investment.php 	modified:   pages/opportunity.php 	modified:   pages/register.php 	modified:   pages/rmmain.php 	modified:   pages/viewinvestment.php

+++ pages/clientmain.php
@@ -29,40 +29,49 @@
          <h2>WMC Client Main</h2>
          <?php
          $infos = GetPreferences($username);
-         ?>
-         <h2>Client Preferences</h2>
-         <table id="preference">
-            <tr>
-               <th align="left" class="p1">Date Submitted</th>
-               <th align="left" class="p1">Asset Type</th>
-               <th align="left" class="p1">Industry Sector</th>
-               <th align="left" class="p1">Country</th>
-               <th align="left" class="p1">Region</th>
-               <th align="left" class="p2">Preference Details</th>
-               <th align="left" class="p3"></th>
-            </tr>
-            <?php
-            foreach ($infos as $info) {
+         if (!empty($infos)) {
             ?>
+            <h2>Client Preferences</h2>
+            <table id="preference">
                <tr>
-                  <td class="p1"><?php echo $info['datesubmitted']; ?></td>
-                  <td class="p1"><?php echo $info['assetdesc']; ?></td>
-                  <td class="p1"><?php echo $info['sectordesc']; ?></td>
-                  <td class="p1"><?php echo $info['countryname']; ?></td>
-                  <td class="p1"><?php echo $info['regionname']; ?></td>
-                  <td class="p2"><?php echo $info['prefdetails']; ?></td>
-                  <td class="p3">
-                     <a href="index.php?page=idea&prefid=<?php echo $info['prefid']; ?>">Extract Idea</a> |
-                     <a href="index.php?page=editpreference&prefid=<?php echo $info['prefid']; ?>">Edit</a> |
-                     <a href="index.php?page=deletepreference&prefid=<?php echo $info['prefid']; ?>"
-                        onclick="return confirm('Are you sure you want to delete this preference?')">Delete</a>
-                  </td>
+                  <th align="left" class="p1">Date Submitted</th>
+                  <th align="left" class="p1">Asset Type</th>
+                  <th align="left" class="p1">Industry Sector</th>
+                  <th align="left" class="p1">Country</th>
+                  <th align="left" class="p1">Region</th>
+                  <th align="left" class="p2">Preference Details</th>
+                  <th align="left" class="p3"></th>
                </tr>
+               <?php
+               foreach ($infos as $info) {
+                  ?>
+                  <tr>
+                     <td class="p1"><?php echo $info['datesubmitted']; ?></td>
+                     <td class="p1"><?php echo $info['assetdesc']; ?></td>
+                     <td class="p1"><?php echo $info['sectordesc']; ?></td>
+                     <td class="p1"><?php echo $info['countryname']; ?></td>
+                     <td class="p1"><?php echo $info['regionname']; ?></td>
+                     <td class="p2"><?php echo $info['prefdetails']; ?></td>
+                     <td class="p3">
+                        <a href="index.php?page=idea&prefid=<?php echo $info['prefid']; ?>">Extract Idea</a> |
+                        <a href="index.php?page=editpreference&prefid=<?php echo $info['prefid']; ?>">Edit</a> |
+                        <a href="index.php?page=deletepreference&prefid=<?php echo $info['prefid']; ?>"
+                           onclick="return confirm('Are you sure you want to delete this preference?')">Delete</a>
+                     </td>
+                  </tr>
+                  <?php
+               }
+               ?>
+            </table>
             <?php
-            }
+         } else{
+            echo "<br>No preferences set as yet!";
             ?>
-         </table>
-
+            <br>
+            <a href="index.php?page=addpreference">Add a preference here!</a>
+            <?php
+         }
+         ?>
       </div>
    </div>
 </body>
